pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "susvibes"
version = "0.1.0"
description = "SusVibes: Benchmarking Vulnerability of Agent-Generated Code in Real-World Tasks"
readme = "README.md"
requires-python = ">=3.11"
license = {file = "LICENSE"}
authors = [
    {name = "SusVibes Team", email = "songwenz@andrew.cmu.edu"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Security",
]

dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.15",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "asttokens==3.0.0",
    "attrs==25.3.0",
    "certifi==2025.7.14",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "comm==0.2.3",
    "debugpy==1.8.11",
    "decorator==5.2.1",
    "distro==1.9.0",
    "docker==7.1.0",
    "exceptiongroup==1.3.0",
    "executing==2.2.0",
    "filelock==3.18.0",
    "frozenlist==1.7.0",
    "fsspec==2025.7.0",
    "h11==0.16.0",
    "hf-xet==1.1.5",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "huggingface-hub==0.34.3",
    "idna==3.10",
    "importlib_metadata==8.7.0",
    "ipykernel==6.30.0",
    "ipython==9.4.0",
    "ipython_pygments_lexers==1.1.1",
    "jedi==0.19.2",
    "Jinja2==3.1.6",
    "jiter==0.10.0",
    "jsonschema==4.25.0",
    "jsonschema-specifications==2025.4.1",
    "jupyter_client==8.6.3",
    "jupyter_core==5.8.1",
    "litellm==1.74.9.post1",
    "MarkupSafe==3.0.2",
    "matplotlib-inline==0.1.7",
    "multidict==6.6.3",
    "nest_asyncio==1.6.0",
    "openai==1.98.0",
    "packaging==25.0",
    "parso==0.8.4",
    "pexpect==4.9.0",
    "pickleshare==0.7.5",
    "platformdirs==4.3.8",
    "prompt_toolkit==3.0.51",
    "propcache==0.3.2",
    "psutil==5.9.0",
    "ptyprocess==0.7.0",
    "pure_eval==0.2.3",
    "pydantic==2.11.7",
    "pydantic_core==2.33.2",
    "Pygments==2.19.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "PyYAML==6.0.2",
    "pyzmq==26.2.0",
    "referencing==0.36.2",
    "regex==2025.7.31",
    "requests==2.32.4",
    "rpds-py==0.26.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "stack_data==0.6.3",
    "tiktoken==0.9.0",
    "tokenizers==0.21.4",
    "tornado==6.5.1",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "typing_extensions==4.14.1",
    "typing-inspection==0.4.1",
    "urllib3==2.5.0",
    "wcwidth==0.2.13",
    "yarl==1.20.1",
    "zipp==3.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1.0",
]

[project.scripts]
susvibes-eval = "susvibes.run_evaluation:cli_main"

[tool.setuptools]
packages = {find = {}}

[tool.setuptools.package-data]
"susvibes" = [
    "safety_strategies/*.yaml",
    "env_specs/*.json",
]

[project.urls]
Homepage = "https://github.com/LeiLiLab/susvibes"
Repository = "https://github.com/LeiLiLab/susvibes"
Issues = "https://github.com/LeiLiLab/susvibes/issues"

